mantle/ore: add ore gcloud update-image
This new command will take advantage of the new Alpha compute API
call to image.patch that will allow us to update the image family
for an image after it has been created. The image.patch API call
also allows for updating the description (which we also allow here)
and the deprecation state (which we leave to the deprecate-image
command).