mantle/kola, mantle/platform: split test buckets if merged config is too large
We have seen in the past that the non-exclusive tests' merged config can
become too large to pass to cloud providers. To fix this issue we compressed
configs (for AWS), but this is only a temporary fix.

If a non-exclusive test bucket's config becomes too large, let's split the
bucket into two buckets to lower the config size.