signal that oscontainer.py and cmd-upload-oscontainer are legacy
These two utilities create old style oscontainer images (i.e.
not runnable containers). They are on the way out in favor of the
new OSTree Native Containers.

This commit renames the files to add `deprecated-legacy-format`
into the filename. It symlinks the old filename back to the new
filename to keep backwards compatiblity. It's a cosmetic change,
but hints at our direction.